fix(fixed-header): animated logo on demand
